news_feed/requirements.txt
feedparser==6.0.10
aiohttp==3.9.1
pydantic==2.5.2
asyncio==3.4.3
pyyaml==6.0.1
python-dotenv==1.0.0
langchain==0.0.345
beautifulsoup4==4.12.2
bs4==0.0.1